refactor: move sync-data-handler related files
